cpuidle: teo: Treat negative sleep lengths as indefinite
A negative sleep length means there's no hrtimer armed, and therefore the
sleep length is indefinite if the tick is stopped. Fix this logic in TEO to
allow a deeper idle state to correctly be used when there's no hrtimer
armed and thus the CPU won't be prematurely woken up from its sleep.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Helium-Studio <67852324+Helium-Studio@users.noreply.github.com>